requirements.txt
streamlit
pandas
numpy
scikit-learn
tensorflow-cpu
nltk
plotly
joblib